ChatWithReActAgent/requirements.txt
langchain
langchain-openai
langchain-google-genai
langchainhub
python-dotenv
streamlit
duckduckgo-search
anthropic
beautifulsoup4
autogenstudio
crewai